Do not change DND if already active.
If the user already has Do Not Disturb active upon entering a screen sharing
or recording session, do not turn it off when it ends.
Use a setting to persist the DND state upon entering a screen sharing or
recording session: a private class member would have lost the state upon
a shell restart or the extension being disabled and reenabled.